requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
attrs==25.1.0
blinker==1.9.0
category_encoders==2.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.2.1
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.6
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.55.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
ipython==8.33.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.7
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.61.0
numpy==1.26.4
nvidia-nccl-cu12==2.23.4
onnx==1.17.0
onnxconverter-common==1.14.0
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.1.4
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
prompt_toolkit==3.0.50
protobuf==3.20.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycaret==3.3.2
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyod==2.0.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.23.1
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
skl2onnx==1.17.0
sklearn-compat==0.1.3
sktime==0.26.0
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
tbats==1.1.3
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wurlitzer==3.1.1
xgboost==2.1.3
xxhash==3.5.0
yellowbrick==1.5
zipp==3.21.0
